No longer set .mediaSource on IMBMediaLibraryParser even if client app cannot handle SSBs
Until now this provided the yellow exclamation mark icon on the library node when the client could not handle SSBs and the library was located outside of a standard directory.

It turns out that it is sufficient to display the yellow exclamation mark icon only on thumbnails in this case (will be able to display thumnails anyhow, only drag&drop is invalidated until user grants access).

This change will also result in much faster library node initialization if client app cannot handle SSBs.