Fix crash 'Identifier of media group (null) must not be nil'
Crash was caused when assertion failed that a media group must not be nil when determining its key photo (and build settings do not void NSAsserts within the release build configuration). Assertion appears to be a bit too much given our graceful handling of the situation once we remove it.

What caused the identifier to not / no longer represent a media group is still somewhat left in the dark. I could not provoke the error even when deleting nodes within the Photos app while the node was already visible in iMedia.