Fix #89, Lightroom library not always accessible when running sandboxed
Had to revert commits 9895e442 and a948b054 because database version cannot be retrieved if the user hasn’t granted access rights to the library’s location yet. And if database version is nil Lightroom library top level node won’t even show up in the tree thus preventing the user to grant access to that library in the first place – you get the point.