Fix issue #92, Loading of "Albums" node of Photos media library takes very long on large libraries
The problem is caused by a potentially very expensive calculation of the so-called „persistentResourceIdentifier“ for an IMBObject if it is in fact an IMBNodeObject (representing a library’s subnode within the browsers IKImageBrowserView).

Since the persistentResoureIdentifiers sole purpose is to conveniently provide an identifier (valid across app sessions) for resources dragged into the client app and IMBNodeObjects (currently) cannot be dragged into the client app anyhow we now just return nil for IMBNodeObjects.

Thanks to Pierre Bernard.